# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.APIKey:
    properties:
      created_at:
        type: string
      description:
        type: string
      expires_at:
        description: NULL = never expires
        type: string
      id:
        type: string
      is_active:
        type: boolean
      key_prefix:
        description: First 8 chars for identification
        type: string
      last_used_at:
        type: string
      name:
        type: string
      scopes:
        description: JSON array of permissions
        items:
          type: integer
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  models.CreateAPIKeyRequest:
    properties:
      description:
        description: API key description
        example: API key for production environment
        type: string
      expires_at:
        description: Optional expiration date (null for no expiration)
        example: "2024-12-31T23:59:59Z"
        type: string
      name:
        description: API key name (3-100 characters)
        example: Production API Key
        maxLength: 100
        minLength: 3
        type: string
      scopes:
        description: List of permission scopes
        example:
        - users:read
        - token:validate
        items:
          type: string
        minItems: 1
        type: array
    required:
    - name
    - scopes
    type: object
  models.CreateAPIKeyResponse:
    properties:
      api_key:
        allOf:
        - $ref: '#/definitions/models.APIKey'
        description: Created API key metadata
      plain_key:
        description: Plain text API key (shown only once at creation)
        example: agw_1234567890abcdef1234567890abcdef
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      details:
        description: Additional error details
        example: Email field is required
        type: string
      error:
        description: HTTP error status text
        example: Bad Request
        type: string
      message:
        description: Human-readable error message
        example: Invalid request parameters
        type: string
    type: object
...
  models.ListAPIKeysResponse:
    properties:
      api_keys:
        description: List of API keys
        items:
          $ref: '#/definitions/models.APIKey'
        type: array
      total:
        description: Total number of API keys
        example: 5
        type: integer
    type: object
...
  models.MessageResponse:
    properties:
      message:
        description: Response message
        example: Operation completed successfully
        type: string
    type: object
...
  models.UpdateAPIKeyRequest:
    properties:
      description:
        description: API key description
        example: Updated description
        type: string
      is_active:
        description: Whether the key is active
        example: true
        type: boolean
      name:
        description: API key name
        example: Updated API Key Name
        type: string
      scopes:
        description: List of permission scopes
        example:
        - users:read
        - users:write
        items:
          type: string
        type: array
    type: object
...
host: localhost:8811
info:
  contact:
    email: maksbalashov@gmail.com
    name: API Support
  description: |-
    Centralized authentication and authorization system for microservices ecosystem.

    ## Overview
    Auth Gateway provides a complete authentication and authorization solution including:
    - User registration and authentication (email/password, OAuth, passwordless)
    - Two-factor authentication (TOTP)
    - API key management for service-to-service communication
    - Role-based access control (RBAC)
    - Session management
    - Webhooks for event notifications

    ## Authentication
    Most endpoints require authentication via JWT Bearer token or API key.

    ### JWT Authentication
    Include the access token in the Authorization header: `Authorization: Bearer {token}`

    ### API Key Authentication
    Include the API key in the X-API-Key header: `X-API-Key: agw_{key}`
  title: Auth Gateway API
  version: "1.0"
paths:
...
  /api/api-keys:
    get:
      description: Get all API keys belonging to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListAPIKeysResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - API Keys
    post:
      consumes:
      - application/json
      description: Create a new API key for service-to-service authentication
      parameters:
      - description: API key data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - API Keys
  /api/api-keys/{id}:
    delete:
      description: Permanently delete an API key
      parameters:
      - description: API Key ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete API key
      tags:
      - API Keys
    get:
      description: Get details of a specific API key owned by the authenticated user
      parameters:
      - description: API Key ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIKey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get API key
      tags:
      - API Keys
    put:
      consumes:
      - application/json
      description: Update API key metadata (name, scopes, etc.)
      parameters:
      - description: API Key ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIKey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update API key
      tags:
      - API Keys
  /api/api-keys/{id}/revoke:
    post:
      description: Revoke an API key to prevent further use
      parameters:
      - description: API Key ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke API key
      tags:
      - API Keys
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'API Key for service-to-service authentication. Format: "agw_{key}"'
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token for user authentication. Format: "Bearer {access_token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"